+++ Bank/src/SecurityAccount.java
@@ -7,7 +7,7 @@ public class SecurityAccount extends Account {
     private double loss;
     private ArrayList<Stock> stockListOwned;
 
-    public SecurityAccount(int accountId, int customerId) {
+    public SecurityAccount(long accountId, int customerId) {
         super(AccountType.SECURITY, accountId, customerId);
         setLoss(0);
         setProfit(0);
